# Added request logger middleware which helps to use custom logger library for logging requests