Added request logger middleware which helps to use custom logger library for logging requests.


Usage would be as follows:

-------

Example for `fmt.Printf`
```go
    e.Use(middleware.RequestLoggerWithConfig(middleware.RequestLoggerConfig{
        LogStatus: true,
        LogURI:    true,
        LogValuesFunc: func(c echo.Context, v middleware.RequestLoggerValues) error {
            fmt.Printf("REQUEST: uri: %v, status: %v\n", v.URI, v.Status)
            return nil
        },
    }))
```

 Example for Zerolog (https://github.com/rs/zerolog)
```go
    logger := zerolog.New(os.Stdout)
    e.Use(middleware.RequestLoggerWithConfig(middleware.RequestLoggerConfig{
        LogURI:    true,
        LogStatus: true,
        LogValuesFunc: func(c echo.Context, v middleware.RequestLoggerValues) error {
            logger.Info().
                Str("URI", v.URI).
                Int("status", v.Status).
                Msg("request")

            return nil
        },
    }))
```

Example for Zap (https://github.com/uber-go/zap)
```go
    logger, _ := zap.NewProduction()
    e.Use(middleware.RequestLoggerWithConfig(middleware.RequestLoggerConfig{
        LogURI:    true,
        LogStatus: true,
        LogValuesFunc: func(c echo.Context, v middleware.RequestLoggerValues) error {
            logger.Info("request",
                zap.String("URI", v.URI),
                zap.Int("status", v.Status),
            )

            return nil
        },
    }))
```

Example for Logrus (https://github.com/sirupsen/logrus)
```go
    log := logrus.New()
    e.Use(middleware.RequestLoggerWithConfig(middleware.RequestLoggerConfig{
        LogURI:    true,
        LogStatus: true,
        LogValuesFunc: func(c echo.Context, values middleware.RequestLoggerValues) error {
            log.WithFields(logrus.Fields{
                "URI":    values.URI,
                "status": values.Status,
            }).Info("request")

            return nil
        },
    }))
```

------



Note about benchmarks for `RequestLogger` - these do not format message and therefore are not 1-to-1 comparable with `Logger` middleware performance. 
```
// without header/query/form fields
BenchmarkLoggerWithConfig_withoutMapFields-6         428011          2713 ns/op        1472 B/op         24 allocs/op
BenchmarkRequestLogger_withoutMapFields-6            721497          1567 ns/op        1384 B/op         17 allocs/op

// with header/query/form fields
BenchmarkLoggerWithConfig_withMapFields-6            330147          3585 ns/op        1937 B/op         29 allocs/op
BenchmarkRequestLogger_withMapFields-6               429079          3113 ns/op        3048 B/op         28 allocs/op
```